Fixed Lack of Stream Issue Due to Failed Array Indices
(I wrote i instead of j)

+++ vision_app/app/src/main/java/com/team254/cheezdroid/NativePart.java
@@ -37,7 +37,7 @@ public class NativePart {
             int s_max,
             int v_min,
             int v_max,
-            int[] out_dis,
+            byte[] out_dis,
             TargetsInfo destInfo);
 
     /**
